F_T30603 move AssessmentTable ValueObjects under core ValueObejcts
- adjust use statements
