+++ demosplan/DemosPlanCoreBundle/Transformers/AssessmentTable/StatementBulkEditTransformer.php
@@ -10,10 +10,10 @@
 
 namespace demosplan\DemosPlanCoreBundle\Transformers\AssessmentTable;
 
-use demosplan\DemosPlanAssessmentTableBundle\ValueObject\StatementBulkEditVO;
 use demosplan\DemosPlanCoreBundle\Logic\ApiRequest\Transformer\BaseTransformer;
 use demosplan\DemosPlanCoreBundle\ResourceTypes\ClaimResourceType;
 use demosplan\DemosPlanCoreBundle\ResourceTypes\StatementResourceType;
+use demosplan\DemosPlanCoreBundle\ValueObject\AssessmentTable\StatementBulkEditVO;
 use demosplan\DemosPlanStatementBundle\Logic\StatementService;
 use demosplan\DemosPlanUserBundle\Logic\UserService;
 use Exception;
