+++ demosplan/DemosPlanStatementBundle/Logic/AssessmentTableExporter/AssessmentTableFileExporterAbstract.php
@@ -12,8 +12,8 @@ declare(strict_types=1);
 
 namespace demosplan\DemosPlanStatementBundle\Logic\AssessmentTableExporter;
 
-use demosplan\DemosPlanAssessmentTableBundle\ValueObject\ExportTemplateData;
 use demosplan\DemosPlanCoreBundle\Logic\AssessmentTable\AssessmentTableServiceOutput;
+use demosplan\DemosPlanCoreBundle\ValueObject\AssessmentTable\ExportTemplateData;
 use demosplan\DemosPlanCoreBundle\ValueObject\ToBy;
 use demosplan\DemosPlanProcedureBundle\Logic\CurrentProcedureService;
 use demosplan\DemosPlanStatementBundle\Logic\AssessmentHandler;
